pyproject.toml
[tool.poetry]
name = "scratch-show"
version = "0.1.0"
description = "Scratch Show is a basic application that uses simple database to work out and srequired"
authors = ["Akshat <akshat.pande@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.31.0"
pandas = "^2.2.0"
requests = "^2.31.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
